优化代码

+++ feat-core/src/main/java/tech/smartboot/feat/core/client/HttpResponse.java
@@ -18,7 +18,7 @@ public interface HttpResponse extends Response {
 
     String getContentType();
 
-    int getContentLength();
+    long getContentLength();
 
     String getCharacterEncoding();
